requirements.txt
eclipse-zenoh==1.6.2
lz4==4.4.4
mcap==1.3.0
numpy==2.2.6
opencv-python==4.12.0.88
zstandard==0.25.0
